sprots-prediction-tool/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
nba_api
textual
requests
scikit-learn
pandas
joblib